+++ packages/serinus/bin/serinus.dart
@@ -4,6 +4,19 @@ import 'dart:io';
 import 'package:serinus/serinus.dart';
 import 'package:shelf/shelf.dart' as shelf;
 
+class TestObj with JsonObject {
+  final String name;
+
+  TestObj(this.name);
+
+  @override
+  Map<String, dynamic> toJson() {
+    return {
+      'name': name,
+    };
+  }
+}
+
 class TestMiddleware extends Middleware {
   int counter = 0;
 
@@ -100,19 +113,16 @@ class PostRoute extends Route {
 class HomeController extends Controller {
   HomeController({super.path = '/'}) {
     on(GetRoute(path: '/'), (context) async {
-      return Response.text('Hello world');
+      return Response.json([
+        TestObj('Hello'),
+        TestObj('World'),
+        {'test': context.query['test']}
+      ]);
     },
         schema: ParseSchema(
-          query: object({
-            'test': string().contains('a'),
-          }),
-          headers: object({
-            'test': string().contains('a'),
-          }),
-          // error: (errors) {
-          //   return BadRequestException(message: 'Invalid query parameters');
-          // }
-        ));
+            query: object({
+          'test': string().encode(),
+        }).optionals(['test'])));
     on(PostRoute(path: '/*'), (context) async {
       return Response.text(
           '${context.request.getData('test')} ${context.params}');