+++ packages/serinus/lib/src/core/middleware.dart
@@ -29,15 +29,19 @@ abstract class Middleware {
   ///
   /// It is used to create a middleware from a shelf middleware giving interoperability between Serinus and Shelf.
   factory Middleware.shelf(Function handler,
-      {List<String> routes = const ['*']}) {
-    return _ShelfMiddleware(handler, routes: routes);
+      {List<String> routes = const ['*'], bool ignoreResponse = true}) {
+    return _ShelfMiddleware(handler,
+        routes: routes, ignoreResponse: ignoreResponse);
   }
 }
 
 class _ShelfMiddleware extends Middleware {
   final dynamic _handler;
 
-  _ShelfMiddleware(this._handler, {super.routes = const ['*']});
+  final bool ignoreResponse;
+
+  _ShelfMiddleware(this._handler,
+      {super.routes = const ['*'], this.ignoreResponse = true});
 
   /// Most of the code has been taken from
   /// https://github.com/codekeyz/pharaoh/tree/main/packages/pharaoh/lib/src/shelf_interop
@@ -71,7 +75,7 @@ class _ShelfMiddleware extends Middleware {
     res.status(response.statusCode);
     res.headers(headers);
     final responseBody = await response.readAsString();
-    if (responseBody.isNotEmpty) {
+    if (responseBody.isNotEmpty && !ignoreResponse) {
       await res.send(utf8.encode(responseBody));
     }
   }